requirements-dev.txt
# Development requirements
matplotlib
cartopy
pytest
pytest-cov
pytest-mpl
coverage
sphinx=1.8.5
sphinx_rtd_theme
sphinx-gallery
numpydoc
